Supplementary Materials/CelebA/requirements.txt
absl-py==1.0.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
cycler==0.11.0
einops==0.3.2
fonttools==4.28.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.0
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
stochman==0.1.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.0.0
tomli==1.2.2
torch==1.10.1
torchplot==0.2.0
torchvision==0.11.2
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
